Addressed Sagar's comments
