+++ server/src/main/java/org/opensearch/common/cache/stats/NoopCacheStatsHolder.java
@@ -17,15 +17,11 @@ import java.util.List;
  * Will be removed once pluggable caches is no longer an experimental feature.
  * Returns all-zero stats when calling getImmutableCacheStatsHolder(). Does keep track of entries for use in ICache.count().
  */
-public class DummyCacheStatsHolder implements CacheStatsHolderInterface {
+public class NoopCacheStatsHolder implements CacheStatsHolder {
     private CounterMetric entries;
-    private CounterMetric sizeInBytes;
-    private final List<String> dimensionNames;
 
-    public DummyCacheStatsHolder(List<String> dimensionNames) {
-        this.dimensionNames = dimensionNames;
+    public NoopCacheStatsHolder() {
         this.entries = new CounterMetric();
-        this.sizeInBytes = new CounterMetric();
     }
 
     @Override
@@ -38,14 +34,10 @@ public class DummyCacheStatsHolder implements CacheStatsHolderInterface {
     public void incrementEvictions(List<String> dimensionValues) {}
 
     @Override
-    public void incrementSizeInBytes(List<String> dimensionValues, long amountBytes) {
-        sizeInBytes.inc(amountBytes);
-    }
+    public void incrementSizeInBytes(List<String> dimensionValues, long amountBytes) {}
 
     @Override
-    public void decrementSizeInBytes(List<String> dimensionValues, long amountBytes) {
-        sizeInBytes.dec(amountBytes);
-    }
+    public void decrementSizeInBytes(List<String> dimensionValues, long amountBytes) {}
 
     @Override
     public void incrementEntries(List<String> dimensionValues) {
@@ -60,7 +52,6 @@ public class DummyCacheStatsHolder implements CacheStatsHolderInterface {
     @Override
     public void reset() {
         this.entries = new CounterMetric();
-        this.sizeInBytes = new CounterMetric();
     }
 
     @Override
@@ -74,6 +65,6 @@ public class DummyCacheStatsHolder implements CacheStatsHolderInterface {
     @Override
     public ImmutableCacheStatsHolder getImmutableCacheStatsHolder() {
         ImmutableCacheStatsHolder.Node dummyNode = new ImmutableCacheStatsHolder.Node("", null, new ImmutableCacheStats(0, 0, 0, 0, 0));
-        return new ImmutableCacheStatsHolder(dummyNode, dimensionNames);
+        return new ImmutableCacheStatsHolder(dummyNode, List.of());
     }
 }
